Move download button output of markdown component
